==> EDA/1_EDA_intro.ipynb <==
# The Academy Awards, 1927 - 2023

Este [dataset](https://www.kaggle.com/datasets/unanimad/the-oscar-award) contiene un scraping de la base de datos de los Premios de la Academia, registros de los nominados y ganadores desde 1927 hasta 2023.

En este EDA cruzaré los datos de diferentes datasets de las películas con las nominadas y ganadoras a mejor pelicula para ver si los miembros de la academia siguen algun tipo de patron o *bias* para elegir la ganadora.

import pandas as pd

df_oscars = pd.read_csv("./data/the_oscar_award.csv")
df_oscars

### Los diferentes nombres que ha tenido la categoria para mejor película a lo largo de los años
1927/28–1928/29: Academy Award for Outstanding Picture  
1929/30–1940: Academy Award for Outstanding Production  
1941–1943: Academy Award for Outstanding Motion Picture  
1944–1961: Academy Award for Best Motion Picture  
1962–present: Academy Award for Best Picture

BEST_PICTURE = df_oscars.category == "BEST PICTURE"
OUTSTANDING_PICTURE =  df_oscars.category == "OUTSTANDING PICTURE"
OUTSTANDING_PRODUCTION = df_oscars.category == "OUTSTANDING PRODUCTION"
OUSTANDING_MOTION_PICTURE = df_oscars.category == "OUTSTANDING MOTION PICTURE"
BEST_MOTION_PICTURE = df_oscars.category == "BEST MOTION PICTURE"

df_best_picture = df_oscars[BEST_PICTURE | OUTSTANDING_PICTURE | OUTSTANDING_PRODUCTION | OUSTANDING_MOTION_PICTURE | BEST_MOTION_PICTURE]
df_best_picture

df_best_picture[df_best_picture.winner]
# para asegurarme de que tengo 95 años de datos

df_best_picture.drop(columns="category", inplace=True) # no la necesito
df_best_picture.info()

|columna|descripcion|
|-|-|
|year_film|Año de estreno|
|year_ceremony| Año en que se celebró la gala|
|ceremony|Número de ceremonia|
|name|Nombre de la productora/los productores|
|film|Título de la película|
|winner|Bool True si ganó|

df_best_picture.sample(50)

df_best_picture.to_csv("./data/best_picture.csv")
