==> EDA/2_EDA_valfa.ipynb <==
import pandas as pd 
import requests
import re
import os
import time
import json

archivo = open("C:/Users/Nur/Documents/tokens/tmdb_token.txt", "r")
contenido = archivo.read()
archivo.close()
token = contenido

# INFO

- [ ] cruzar con tmdb, rotten tomatoes, etc
- [ ] oscar bait
- [ ] [IMDB academy awards](https://www.imdb.com/event/ev0000003/2023/1/?ref_=ev_eh)
- [ ] [Cinema timeline](https://www.timetoast.com/timelines/evolution-of-movies-75835972-1771-4afa-8a8c-6b3356decbf6)
- [ ] [AI Movie Companion for Ratings](https://routinehub.co/shortcut/15932/)

# PRUEBAS

## API

[TMDB API](https://developer.themoviedb.org/reference/search-movie)

url = "https://api.themoviedb.org/3/search/movie?query=wings&include_adult=false&language=en-US&page=1&year=1927"

headers = {
    "accept": "application/json",
    "Authorization": f"Bearer {token}"
}

response = requests.get(url, headers=headers)

print(response.text)

genre_ids  
original language  


url = "https://api.themoviedb.org/3/genre/movie/list?language=en"

headers = {
    "accept": "application/json",
    "Authorization": f"Bearer {token}"
}

response = requests.get(url, headers=headers)

genres = response.json()

with open('./data/genres.txt', 'w') as file:
        json.dump(genres, file, indent=2)

with open("./data/genres.json", 'w') as g:
    json.dump(genres, g, indent=2)

df_genres = pd.DataFrame(genres["genres"])
df_genres

df_genres.to_csv("./data/genres.csv")

## REQUEST

df_best_picture = pd.read_csv("./data/best_picture.csv", index_col=0)
df_best_picture.sample(1)

df_prueba = df_best_picture.loc[7288]
df_prueba = pd.DataFrame(df_prueba).T
df_prueba

# movie_data = []
# for i in df_best_picture.loc[7288]:
#     # Enviar una petición a la API
#     name = df_best_picture["film"].str.replace(" ", "%20")
#     year = df_best_picture["year_film"]
#     url = f"https://api.themoviedb.org/3/search/movie?query={name}&include_adult=false&language=en-US&page=1&year={year}"
#     headers = {
#     "accept": "application/json",
#     "Authorization": f"Bearer {token}"
#     }

#     response = requests.get(url, headers=headers)

#     # Esperar a que se complete la petición
#     time.sleep(1/50)

#     movie_data.append(response.json())
# movie_data

movie_data = []
base_url = "https://api.themoviedb.org/3/search/movie?"

for index, row in df_best_picture.iterrows():
    # Obtener el nombre de la película y el año de la fila actual
    name = row['film'].replace(" ", "%20")
    year = row['year_film']

    # Construir la URL con los parámetros de la búsqueda
    url = f"{base_url}query={name}&include_adult=false&language=en-US&page=1&year={year}"

    # Configurar los encabezados con el token de autenticación
    headers = {
        "accept": "application/json",
        "Authorization": f"Bearer {token}"
    }

    # Realizar la solicitud a la API
    response = requests.get(url, headers=headers)
    time.sleep(1/50)
    
    # Que el titulo del resultado sea exactamente el mismo q el del df
    for result in response.json()['results']:
        if result['title'] == row['film']:
            movie_data.append(result)

# Ahora movie_data contiene la información de todas las películas en df_best_picture


df_movie_data = pd.DataFrame(movie_data)
df_movie_data.sample(1)

df_movie_data.drop(columns=["adult", "backdrop_path", "poster_path", "video"], inplace=True)
df_movie_data

df_movie_data.to_csv("./data/movie_data.csv")

## JSON


for i in range(len(df_best_picture)):
    url = "https://api.themoviedb.org/3/search/movie?query=wings&include_adult=false&language=en-US&page=1&year=1927"

    headers = {
    "accept": "application/json",
    "Authorization": "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJkODZjMWUxNWRhMDI2Y2E0NDZhMzUyYTMwODY4N2ZjYyIsInN1YiI6IjY1NjIyOTc5MzY3OWExMDk3OTA1ZGRlNyIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.ljIozDoGs4w7O-LXH53YCTgJyvvKk2zW3M4wJMG0O5g"
    }

response = requests.get(url, headers=headers)

print(response.json())

response.json()

datos = [json.loads(linea) for linea in f]
print(type(datos))
datos
