==> EDA/3_dfs.ipynb <==
import pandas as pd 
import requests
import re
import os
import time
import json

archivo = open("C:/Users/Nur/Documents/tokens/tmdb_token.txt", "r")
contenido = archivo.read()
archivo.close()
token = contenido

df_best_picture = pd.read_csv("./data/best_picture.csv", index_col=0)
df_movie_data = pd.read_csv("./data/movie_data.csv", index_col=0)
df_genres = pd.read_csv("./data/genres.csv", index_col=0)

df_movie_data.info()

df_best_picture.info()

# me faltan pelis

df_best_picture.sample(1)

df_movie_data.sample(1)

df_genres.sample(1)

with open("./data/genres.txt", 'r') as f:
    genres = f.read()

with open("./data/genres.json", 'r') as f:
    genres = json.load(f)

genres["genres"]

# GPT
# diccionario_resultante = {item['id']: item['name'] for item in lista_diccionarios}

dict_genres = {}

for i in genres["genres"]:
    id = i["id"]
    name = i["name"]
    dict_genres[id] = name

dict_genres

df_movie_data['genre'] = df_movie_data['genre_ids'].apply(dict_genres.get)

df_movie_data["genre_ids"].loc[0][]

df_prueba = pd.DataFrame(df_movie_data.loc[0]).T
df_prueba

# reemplazar los ids con los nombres de los generos
for index, row in df_prueba.iterrows():
    string = row["genre_ids"]
    # convertir string a lista
    # reemplazar cada id con su nombre

import ast
for index, row in df_movie_data.iterrows():
    string = row["genre_ids"]
    
    # Convertir la string a una lista
    lista_ids = ast.literal_eval(string)
    
    # Mapear cada ID a su nombre de género
    genre_nombres = [dict_genres.get(i) for i in lista_ids]
    
    # Asignar la lista resultante a la columna 'genre'
    df_movie_data.at[index, "genre"] = genre_nombres

df_movie_data

# ??? a la hora de analizar tener varios generos en la misma columna sera un problema
# a tener en cuenta

df_movie_data.info()

df_pelis = pd.merge(df_best_picture, df_movie_data, left_on='film', right_on='title', how="left")
df_pelis

df_pelis.info()

df_pelis[df_pelis["film"].duplicated()]

# ??? liada, tengo q fijarme en film/title/original title

url = "https://api.themoviedb.org/3/search/movie?query=taxi%20driver&include_adult=false&language=en-US&page=1&year=1976"

headers = {
    "accept": "application/json",
    "Authorization": f"Bearer {token}"
}

response = requests.get(url, headers=headers)

taxi_driver = response.json()

taxi_driver
