==> EDA/5.ipynb <==
import pandas as pd

Arreglar el df ratings

df_ratings = pd.read_csv("./data/ratings.csv")
df_ratings.drop(columns="Unnamed: 0", inplace=True)

df_ratings.loc[153]

df_ratings[df_ratings.duplicated(['title', 'rating'])]

df_ratings = df_ratings.drop(154)

# Restablecer el índice después de eliminar la fila
df_ratings = df_ratings.reset_index(drop=True)

df_ratings[df_ratings.duplicated(['title', 'rating'])]

df_ratings

df_ratings.to_csv("./data/ratings.csv", index=False)
